Removed the saving of "output" within the scaffold.
Using execSync to run a command and using "inherit" will put the child
command output into the parent.  The problem is that it's no longer
returned by execSync, so it's not possible to save it within the output
variable (it's null).  Better to have it update the terminal when it
runs within the parent than to save it and print when over.  Long
running commands will update the terminal as they run.